+++ apps/web/.env
@@ -1,2 +1,4 @@
 SKIP_PREFLIGHT_CHECK=true
 HTTPS=true
+BROWSER=none
+CI=false
