+++ apps/web/.eslintrc.js
@@ -4,6 +4,7 @@ module.exports = {
   rules: {
     'react-hooks/rules-of-hooks': 'error',
     'react-hooks/exhaustive-deps': 'warn',
+    'prefer-template': 'error',
   },
   settings: {
     react: {
