+++ apps/web/src/components/Footer.tsx
@@ -15,10 +15,10 @@ const LinkStyled = styled('a')(({ theme }) => ({
 
 export default function Footer() {
   const domain = '.no'
-  const improvementsEmail =
-    'dataplattform' + String.fromCharCode(64) + `knowit${domain}`
-  const datacorrectionEmail =
-    'drift' + String.fromCharCode(64) + `knowit${domain}`
+  const improvementsEmail = `dataplattform${String.fromCharCode(
+    64
+  )}knowit${domain}`
+  const datacorrectionEmail = `drift${String.fromCharCode(64)}knowit${domain}`
 
   const improvementsHref = `mailto:${improvementsEmail}`
   const correctionHref = `mailto:${datacorrectionEmail}?subject=Melding%20om%20datafeil&body=Feilen%20ble%20funnet%20her%3A%20%0D%0AOg%20omhandler%20dette%3A%20`
