+++ apps/web/src/components/gridItem/GridItemContent.tsx
@@ -1,7 +1,6 @@
-import React from 'react'
 import { styled } from '@mui/material/styles'
 
-const ComponentRoot = styled('div')(({ theme }) => ({
+export const GridItemContent = styled('div')(({ theme }) => ({
   width: '100%',
   padding: 15,
   fontSize: 16,
@@ -11,11 +10,3 @@ const ComponentRoot = styled('div')(({ theme }) => ({
   borderRight: `1px solid ${theme.palette.background.darker}`,
   borderRadius: '0px 0px 6px 6px',
 }))
-
-interface GridItemContentProps {
-  children: React.ReactNode | React.ReactNode[]
-}
-
-export function GridItemContent({ children }: GridItemContentProps) {
-  return <ComponentRoot>{children}</ComponentRoot>
-}
