+++ apps/web/src/components/gridItem/GridItemHeader.tsx
@@ -31,6 +31,7 @@ const GridHeaderTitle = styled('h2', {
     fontWeight: green ? 'bold' : big ? 'normal' : 700,
     paddingLeft: big && 11,
     color: green && '#FFFFFF',
+    margin: 0,
   })
 )
 
@@ -69,7 +70,7 @@ export function GridItemHeader({
       <Grid container direction="row" alignItems="center">
         <GridHeaderTitle big={big} green={green} longTitleText={longTitleText}>
           {card && clickable ? (
-            <OpenInNewLink to={'/kunder/' + title} target="_blank">
+            <OpenInNewLink to={`/kunder/${title}`} target="_blank">
               {title}
             </OpenInNewLink>
           ) : (
