+++ apps/web/src/components/sortableTable/components/CvDialog.tsx
@@ -278,7 +278,7 @@ export function createXlsLinks(rows: string[][], filename: string) {
     const url = window.URL.createObjectURL(blob)
     const link = document.createElement('a')
     link.href = url
-    link.download = filename + '.xlsx'
+    link.download = `${filename}.xlsx`
     link.click()
   })
 }
