+++ apps/web/src/index.tsx
@@ -1,4 +1,3 @@
-import React from 'react'
 import { BrowserRouter } from 'react-router-dom'
 
 import App from './App'
