+++ apps/web/src/pages/customer/CustomerPage.tsx
@@ -1,7 +1,7 @@
 import React from 'react'
 import NavTab from '../../components/header/NavTab'
-import CustomerList from './components/CustomerList'
-import { CustomerOverview } from './components/CustomerOverview'
+import CustomerList from './customer-list/CustomerList'
+import { CustomerOverview } from './customer-overview/CustomerOverview'
 import { pageTitle } from '../../utils/pagetitle'
 
 export default function Customer() {
