--- apps/web/src/pages/customer/cards/index.ts
@@ -1,2 +1,0 @@
-export { default as HoursBilledPerCustomerCard } from './HoursBilledPerCustomerCard'
-export { default as HoursBilledPerWeekCard } from './HoursBilledPerWeekCard'
