+++ apps/web/src/pages/customer/customer-overview/CustomerOverview.tsx
@@ -1,30 +1,20 @@
 import React, { useEffect, useState } from 'react'
 import CustomerCardListOverview from './CustomerCardListOverview'
 import { Grid } from '@mui/material'
-import { HoursBilledPerWeekCard } from '../cards'
-import { useCustomerCards } from '../../../api/data/customer/customerQueries'
 import storageTokens from '../util/local-storage-tokens'
-import { useMatomo } from '@jonkoops/matomo-tracker-react'
+import CustomerHoursPerWeekSection from './CustomerHoursPerWeekSection'
+import { useSelectedCustomerIds } from '../util/local-storage-hooks'
 
 export const CustomerOverview = () => {
-  const { data } = useCustomerCards()
-  const customers = data?.map((item) => item.customer)
-  const [showHistoricCustomer, setShowHistoricCustomer] = useState(false)
-  const [selectedCustomerIds, setSelectedCustomerIds] = useState(null)
+  const [showHistoricCustomers, setShowHistoricCustomers] = useState(false)
+  const { selectedCustomerIds, setSelectedCustomerIds } =
+    useSelectedCustomerIds()
   const [selectedPeriodStartDate, setPeriodStartDate] = useState(null)
   const [selectedPeriodEndDate, setPeriodEndDate] = useState(null)
-  const { trackEvent } = useMatomo()
 
   useEffect(() => {
-    const selectedCustomerIds = localStorage.getItem('selectedCustomerIds')
     const startDate = storageTokens.getPeriodStartDate()
     const endDate = storageTokens.getPeriodEndDate()
-
-    if (selectedCustomerIds) {
-      setSelectedCustomerIds(JSON.parse(selectedCustomerIds))
-    } else {
-      setSelectedCustomerIds([])
-    }
 
     setPeriodStartDate(startDate)
     setPeriodEndDate(endDate)
@@ -46,17 +36,12 @@ export const CustomerOverview = () => {
     storageTokens.setPeriodEndDate(selectedPeriodEndDate)
   }, [selectedPeriodEndDate])
 
-  const handleCustomerHistory = () => {
-    setShowHistoricCustomer(!showHistoricCustomer)
-  }
-
   const handleCheckboxChange = (
     event: React.ChangeEvent<HTMLInputElement>,
     customerId: string
   ) => {
-    trackEvent({ category: 'filter-kunde', action: 'click-event' })
     if (event.target.checked) {
-      setSelectedCustomerIds([...selectedCustomerIds, customerId])
+      setSelectedCustomerIds([...selectedCustomerIds, customerId].sort())
     } else {
       setSelectedCustomerIds(
         selectedCustomerIds.filter((id) => id !== customerId)
@@ -66,12 +51,9 @@ export const CustomerOverview = () => {
 
   return (
     <Grid container spacing={2}>
-      <HoursBilledPerWeekCard
+      <CustomerHoursPerWeekSection
         selectedCustomerIds={selectedCustomerIds}
-        setSelectedCustomerIds={setSelectedCustomerIds}
         handleCheckboxChange={handleCheckboxChange}
-        selectedPeriodStartDate={selectedPeriodStartDate}
-        selectedPeriodEndDate={selectedPeriodEndDate}
         handleDateRangeChange={function (
           startDate?: Date,
           endDate?: Date
@@ -79,15 +61,15 @@ export const CustomerOverview = () => {
           setPeriodStartDate(startDate)
           setPeriodEndDate(endDate)
         }}
-        customersWithConsultants={customers}
-        customerHistory={showHistoricCustomer}
-        handleCustomerHistory={handleCustomerHistory}
-        customerSpecificGraph={false}
+        selectedPeriodStartDate={selectedPeriodStartDate}
+        selectedPeriodEndDate={selectedPeriodEndDate}
+        setSelectedCustomerIds={setSelectedCustomerIds}
+        showCustomerHistory={showHistoricCustomers}
+        setShowCustomerHistory={setShowHistoricCustomers}
       />
       <CustomerCardListOverview
         selectedCustomerIds={selectedCustomerIds}
-        showHistoricalData={showHistoricCustomer}
-        customersWithConsultants={customers}
+        showHistoricalData={showHistoricCustomers}
         handleCheckboxChange={handleCheckboxChange}
       />
     </Grid>
