+++ apps/web/src/utils/pagetitle.ts
@@ -1,3 +1,3 @@
 export const pageTitle = (subtitle: string) => {
-  document.title = subtitle + ' - Knowit Folk'
+  document.title = `${subtitle} - Knowit Folk`
 }
